Lab5/requirements.txt
azure-core==1.14.0
azure-eventhub==5.5.0
cassandra-driver==3.25.0
redis==3.5.3
